+++ src/cmd/ksh93/sh/macro.c
@@ -966,9 +966,9 @@ static void mac_substitute(Mac_t *mp, char *cp,char *str,ssize_t subexp[],ssize_
 				continue;
 			}
 			n -= '0';
-			if(n<subsize && (c=(ptrdiff_t)subexp[2*n])>=0)
+			if(n<subsize && (c=subexp[2*n])>=0)
 			{
-				if((n=(ptrdiff_t)subexp[2*n+1]-c)>0)
+				if((n=subexp[2*n+1]-c)>0)
 					mac_copy(mp,str+c,n);
 			}
 		}
@@ -1282,7 +1282,7 @@ retry1:
 		else if(sh.cur_line)
 		{
 			sh.used_pos = 1;
-			v = getdolarg((ptrdiff_t)c,&vsize);
+			v = getdolarg(c,&vsize);
 		}
 #endif  /* SHOPT_FILESCAN */
 		else if(c <= sh.st.dolc)
@@ -1958,7 +1958,7 @@ retry2:
 					{
 						vlast = v;
 						vsize_last = vsize;
-						vsize = (ptrdiff_t)match[0];
+						vsize = match[0];
 					}
 					else if(c=='#')
 						vsize = 0;
@@ -2392,7 +2392,7 @@ static void comsubst(Mac_t *mp,Shnode_t* t, char type)
 			lastc = str[c];
 			str[c] = 0;
 		}
-		mac_copy(mp,str,(ptrdiff_t)c);
+		mac_copy(mp,str,c);
 	}
 	if(was_interactive)
 		sh_onstate(SH_INTERACTIVE);
